+++ src/app/plugin-custom-state.ts
@@ -2,6 +2,7 @@ import { SymmetryOperator } from 'molstar/lib/mol-math/geometry';
 import { Mat4 } from 'molstar/lib/mol-math/linear-algebra';
 import { PluginContext } from 'molstar/lib/mol-plugin/context';
 import { StateSelection, StateTransform } from 'molstar/lib/mol-state';
+import { JSXElementConstructor } from 'react';
 import { Subject } from 'rxjs';
 import { InitParams } from './spec';
 
@@ -46,9 +47,15 @@ export interface PluginCustomState {
             rmsds: string[][],
             coordinateSystems: (SymmetryOperator | undefined)[],
         },
-
     },
     superpositionError?: string,
+    extensions?: {
+        [extensionId: string]: {} | undefined,
+    },
+    customControls?: {
+        viewportTopCenter?: Map<string, JSXElementConstructor<{}>>,
+        viewportTopLeft?: Map<string, JSXElementConstructor<{}>>,
+    },
 }
 
 export interface ClusterMember { pdb_id: string, auth_asym_id: string, struct_asym_id: string, entity_id: number, is_representative: boolean };
@@ -59,4 +66,22 @@ export interface Segment { segment_start: number, segment_end: number, clusters:
  * Supports getting and setting properties. */
 export function PluginCustomState(plugin: PluginContext): PluginCustomState {
     return (plugin.customState as any) ??= {};
+}
+
+export function getExtensionCustomState<T extends {}>(plugin: PluginContext, extensionId: string): Partial<T> {
+    const extensionStates = PluginCustomState(plugin).extensions ??= {};
+    const extensionState: Partial<T> = extensionStates[extensionId] ??= {};
+    return extensionState;
+}
+export function clearExtensionCustomState(plugin: PluginContext, extensionId: string): void {
+    const extensionStates = PluginCustomState(plugin).extensions ??= {};
+    delete extensionStates[extensionId];
+}
+export function extensionCustomStateGetter<StateType extends {}>(extensionId: string) {
+    return (plugin: PluginContext) => getExtensionCustomState<StateType>(plugin, extensionId);
+}
+
+export function CustomControls(plugin: PluginContext, region: keyof NonNullable<PluginCustomState['customControls']>) {
+    const customControls = PluginCustomState(plugin).customControls ??= {};
+    return customControls[region] ??= new Map<string, JSXElementConstructor<{}>>();
 }